@@ -1,12 +1,11 @@
 import { Component, Input, ViewChild, ElementRef, EventEmitter } from '@angular/core';
 import { MapService } from '../services/map.service';
-
-var Lealflet = require('leaflet');
+import { } from 'leaflet';
 
 declare var L: any;
 
 @Component({
-  moduleId: module.id,
+  moduleId: module.id.toString(),
   selector: 'leaf-element',
   templateUrl: 'map.html',
   styleUrls: ['map.css'],
@@ -16,10 +15,13 @@ declare var L: any;
 export class LeafletElement {
   @Input() lat: number = 52.6;
   @Input() lon: number = -1.1;
+  @Input() x: number;
+  @Input() y: number;
   @Input() zoom: number = 12;
   @Input() minZoom: number = 4;
   @Input() maxZoom: number = 19;
   @Input() layerControl: boolean = false;
+  @Input() crs: any = L.CRS.EPSG3857;
   @ViewChild('map') mapElement: ElementRef;
 
   layerControlObject = null;
@@ -28,14 +30,35 @@ export class LeafletElement {
   }
 
   ngOnInit() {
+
+    if (this.x !== undefined) {
+      this.lon = this.x;
+    }
+
+    if (this.y !== undefined) {
+      this.lat = this.y;
+    }
+
+    if (typeof (this.crs) === "string") {
+      var splitCrs = this.crs.split(".");
+      if (splitCrs[0] === "L") {
+        this.crs = L[splitCrs[1]][splitCrs[2]];
+      } else {
+        console.warn("something is not right, reverting to default EPSG3857");
+        this.crs = L.CRS.EPSG3857;
+      }
+    }
+
     let map = L.map(this.mapElement.nativeElement, {
+      crs: this.crs,
       zoomControl: false,
       center: L.latLng(this.lat, this.lon),
       zoom: this.zoom,
       minZoom: this.minZoom,
       maxZoom: this.maxZoom,
       layers: [],
-      closePopupOnClick: false
+      closePopupOnClick: false,
+      continuousWorld: false,
     });
     this.mapElement.nativeElement.myMapProperty = map;
 
